fix issue with disabled schedule exec option
