+++ app/packages/operators/src/SplitButton.tsx
@@ -15,7 +15,7 @@ import {
 } from "@mui/material";
 import ArrowDropDownIcon from "@mui/icons-material/ArrowDropDown";
 import { onEnter } from "./utils";
-import { useTheme } from "@fiftyone/components";
+import ExecutionOptionItem from "./ExecutionOptionItem";
 
 const ButtonStylesOverrides: ButtonProps["sx"] = {
   color: (theme) => theme.palette.text.secondary,
@@ -63,6 +63,7 @@ export default function SplitButton({
   };
 
   const handleSelect = (option) => {
+    if (!option.onSelect) return;
     option.onSelect();
     setOpen(false);
   };
@@ -142,7 +143,7 @@ export default function SplitButton({
                                 : theme.palette.text.disabled,
                           }}
                           primary={
-                            <PrimaryWithTag
+                            <ExecutionOptionItem
                               label={option.choiceLabel || option.label}
                               tag={option.tag}
                               disabled={option.disabled || !option.onClick}
@@ -169,26 +170,5 @@ export default function SplitButton({
         </Popper>
       )}
     </React.Fragment>
-  );
-}
-
-function PrimaryWithTag({ label, tag, disabled }) {
-  const theme = useTheme();
-  const tagEl = tag ? (
-    <span
-      style={{
-        fontSize: "11px",
-        color: disabled ? theme.text.secondary : theme.custom.primarySoft,
-        marginLeft: "5px",
-      }}
-    >
-      {tag}
-    </span>
-  ) : null;
-  return (
-    <div style={{ display: "flex", alignItems: "center" }}>
-      {label}
-      {tagEl}
-    </div>
   );
 }
