+++ app/packages/operators/src/state.ts
@@ -246,6 +246,7 @@ export type OperatorExecutionOption = {
   default?: boolean;
   selected?: boolean;
   onSelect?: () => void;
+  isDisabledSchedule?: boolean;
 };
 
 const useOperatorPromptSubmitOptions = (
@@ -346,6 +347,7 @@ const useOperatorPromptSubmitOptions = (
       id: "disabled-schedule",
       description: markdownDesc,
       isDelegated: true,
+      isDisabledSchedule: true,
     });
   }
